The most significant changes include the refactoring of the `WindController` class in `WindController_PG.h` to gradually adjust wind direction and speed, and the addition of methods to get a string representation of the wind direction and a `sf::VertexArray` for the wind direction indicator. The `ship` member variable in `World_PG.h` has been renamed to `playerShip` for clarity. The `setUpWorld` method in `World_PG.cpp` has been updated to set up the player ship as a Frigate instead of a Galleon. The `gameLoop` method in `World_PG.cpp` and `World_Plat.cpp` have been updated for better clarity and precision.
Changes:

1. `drawGrid` and `getRandomPositions` methods in `Map.h` and `Map.cpp` updated to be `const` methods.
2. `getRandomPosition` method in `Map.h` and `Map.cpp` updated to use `std::uniform_real_distribution<float>`.
3. `getFireKeyString`, `getManualAimKeyString`, `getPortMouseButtonString`, and `getStarboardMouseButtonString` methods in `PlayerShipInputHandler_PG.h` updated to be `const` methods.
4. `WindController` class in `WindController_PG.h` significantly refactored.
5. `ship` member variable in `World_PG.h` renamed to `playerShip`.
6. `setInteractablePositions` method in `IslandMenu_PG.cpp` updated to use `3.5f` for padding.
7. `updateVelocity` method in `ShipMovementHandler_PG.cpp` updated to gradually decrease speed.
8. `setUpWorld` method in `World_PG.cpp` updated to set up player ship as a Frigate.
9. `gameLoop` method in `World_PG.cpp` updated to use `playerShip`.
10. `addInteractablesToMenu` and `updateShipPropertiesString` methods in `inGameHUD_PG.cpp` updated to add and update a display for the wind direction.
11. `setInteractablePositions` method in `inGameHUD_PG.cpp` updated to create a wind direction indicator using a circle and text.
12. `draw` method in `inGameHUD_PG.cpp` updated to draw the wind direction indicator as a circle and text.
13. `gameLoop` method in `World_Plat.cpp` updated to explicitly cast the window size to float when setting the view size.

+++ Platformer Game/Resource/World_Plat.cpp
@@ -19,7 +19,7 @@ void World::gameLoop(sf::Event event) {
 
 	map.drawGrid(window);
 	view.setCenter(player.getPosition());
-	view.setSize(window.getSize().x, window.getSize().y);
+	view.setSize(static_cast<float>(window.getSize().x), static_cast<float>(window.getSize().y));
 	window.setView(view);
 }
